Some fixes after rebasing
