+++ src/modules/navigation/AuthStack.js
@@ -1,10 +1,10 @@
 import { createStackNavigator } from 'react-navigation-stack';
-import LoginScreen from '../../views/loginScreen/LoginScreen';
 import { ksStyle } from '../../styles/basic/ksBasic';
 import AuthCard from '../../components/login/authCard';
 import EditProfileCard from '../../components/login/editProfileCard';
 import MsgAlert from '../../components/login/msgAlert/MsgAlert';
 import ErrorOverlay from '../../components/login/loginForm/ErrorOverlay';
+import LoginScreen from '../../screens/login/LoginScreen';
 
 const AuthStack = createStackNavigator(
   {
