+++ src/screens/drawerNavigator/LogoutView.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1.0,
+    backgroundColor: 'red',
+  },
+});
+
+export default () => {
+  return (
+    <View style={styles.container}>
+      <Text>Logout</Text>
+    </View>
+  );
+};
